Relocate apps.cluster to container_service.clusters
