+++ bcs-app/backend/components/paas_cc.py
@@ -19,9 +19,9 @@ from typing import Dict, List
 from django.conf import settings
 from django.utils.translation import ugettext_lazy as _
 
-from backend.apps.cluster.models import CommonStatus
 from backend.components.base import BaseHttpClient, BkApiClient, ComponentAuth, response_handler
 from backend.components.utils import http_delete, http_get, http_patch, http_post, http_put
+from backend.container_service.clusters.models import CommonStatus
 from backend.utils.basic import getitems
 from backend.utils.decorators import parse_response_data
 from backend.utils.errcodes import ErrorCode
