+++ bcs-app/backend/container_service/clusters/models.py
@@ -142,7 +142,7 @@ class GcloudPollingTask(models.Model):
 
     def polling_task(self):
         """轮训任务"""
-        from backend.apps.cluster import tasks
+        from backend.container_service.clusters import tasks
 
         tasks.ClusterOrNodeTaskPoller.start(
             {"model_type": self.__class__.__name__, "pk": self.pk}, tasks.TaskStatusResultHandler
