+++ bcs-app/backend/container_service/clusters/node_tasks.py
@@ -17,10 +17,10 @@ from celery import shared_task
 from django.utils.translation import ugettext_lazy as _
 
 from backend.apps import constants
-from backend.apps.cluster import models
-from backend.apps.cluster.constants import K8S_SKIP_NS_LIST
 from backend.components import paas_cc
 from backend.components.bcs import k8s, mesos
+from backend.container_service.clusters import models
+from backend.container_service.clusters.constants import K8S_SKIP_NS_LIST
 from backend.utils.errcodes import ErrorCode
