+++ bcs-app/backend/container_service/clusters/serializers.py
@@ -20,10 +20,10 @@ from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
 
 from backend.apps import constants
-from backend.apps.cluster import constants as cluster_constants
-from backend.apps.cluster.models import ClusterInstallLog, NodeLabel, NodeStatus, NodeUpdateLog
 from backend.components import data as data_api
 from backend.components import paas_cc
+from backend.container_service.clusters import constants as cluster_constants
+from backend.container_service.clusters.models import ClusterInstallLog, NodeLabel, NodeStatus, NodeUpdateLog
 from backend.utils.errcodes import ErrorCode
 from backend.utils.error_codes import error_codes
 from backend.utils.exceptions import ResNotFoundError
