+++ bcs-app/backend/container_service/clusters/views/area.py
@@ -17,8 +17,8 @@ from rest_framework import viewsets
 from rest_framework.renderers import BrowsableAPIRenderer
 from rest_framework.response import Response
 
-from backend.apps.cluster.views.utils import get_areas
 from backend.components import paas_cc
+from backend.container_service.clusters.views.utils import get_areas
 from backend.utils.errcodes import ErrorCode
 from backend.utils.error_codes import error_codes
 from backend.utils.renderers import BKAPIRenderer
