+++ bcs-app/backend/container_service/clusters/views/metric_handler/namespace.py
