+++ bcs-app/backend/container_service/clusters/views/node_views/base.py
